backend/requirements.txt
flask==3.0.3
flask-socketio==5.3.6
flask-login==0.6.3
flask-cors==4.0.1
bcrypt==4.1.3
requests==2.31.0
python-dotenv==1.0.1
pywin32==306
psutil==5.9.8
reportlab==4.1.0
eventlet==0.35.2
